refactor: share the untyped-AST child walk between the closure scanners
CapturedVariableScanner and typing.ReturnNodeDetector each carried an
identical 100-line visitChildren match over AST.Node. It now lives once in
ASTWalk, and both scans delegate to it (243+144 lines -> 137+39+118).
